backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-WTF==1.2.1
Flask-Limiter==3.3.0
python-dotenv==1.0.0